# backend/Board/src/main/java/Backend/Board/dto/TaskDTO.java
package Backend.Board.dto;

import com.fasterxml.jackson.annotation.JsonInclude;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.List;

@Data
@NoArgsConstructor
@JsonInclude(JsonInclude.Include.NON_NULL)
public class TaskDTO {
    private Long id;
    
    @NotBlank(message = "Task title is required")
    @Size(min = 1, max = 200, message = "Task title must be between 1 and 200 characters")
    private String title;
    
    @Size(max = 1000, message = "Task description must not exceed 1000 characters")
    private String description;
    
    private String priority;
    
    private String status;
    
    private LocalDateTime dueDate;
    
    private LocalDateTime createdAt;
    
    private LocalDateTime updatedAt;
    
    private List<CommentDTO> comments;
    
    private List<AttachmentDTO> attachments;
    
    private List<LabelDTO> labels;
    
    private UserDTO createdBy;
    
    private UserDTO assignee;
    
    // Constructor for task preview (minimal data)
    public TaskDTO(Long id, String title) {
        this.id = id;
        this.title = title;
    }
    
    // Constructor for task creation
    public TaskDTO(String title, String description, String priority, String status, LocalDateTime dueDate) {
        this.title = title;
        this.description = description;
        this.priority = priority;
        this.status = status;
        this.dueDate = dueDate;
    }
    
    // Constructor for full task response
    public TaskDTO(Long id, String title, String description, String priority, String status, 
                   LocalDateTime dueDate, LocalDateTime createdAt, LocalDateTime updatedAt,
                   List<CommentDTO> comments, List<AttachmentDTO> attachments, List<LabelDTO> labels,
                   UserDTO createdBy, UserDTO assignee) {
        this.id = id;
        this.title = title;
        this.description = description;
        this.priority = priority;
        this.status = status;
        this.dueDate = dueDate;
        this.createdAt = createdAt;
        this.updatedAt = updatedAt;
        this.comments = comments;
        this.attachments = attachments;
        this.labels = labels;
        this.createdBy = createdBy;
        this.assignee = assignee;
    }
}