# frontend/kanban-board-client-app/swagger/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kanban Board API",
    "description": "Advanced Kanban Board API with search, analytics, performance optimization, and comprehensive features. This API exposes only DTOs for clean separation between internal models and API contracts.",
    "contact": {
      "name": "Kanban Board Team",
      "url": "https://kanbanboard.com",
      "email": "support@kanbanboard.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.kanbanboard.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/performance/batch/users": {
      "post": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "getUsersBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/performance/batch/tasks": {
      "post": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "getTasksBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/performance/batch/boards": {
      "post": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "getBoardsBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/performance/async/tasks": {
      "post": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "getTasksAsync",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/performance/cache/user/{userId}": {
      "get": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "getCachedUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "evictUserCache",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/performance/cache/task/{taskId}": {
      "get": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "getCachedTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "evictTaskCache",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/performance/cache/board/{boardId}": {
      "get": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "getCachedBoard",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "evictBoardCache",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/performance/async/board/{boardId}": {
      "get": {
        "tags": [
          "performance-controller"
        ],
        "operationId": "getBoardAsync",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RoleDTO": {
        "type": "object",
        "description": "Role data transfer object for API communication",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the role",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the role",
            "example": "ADMIN"
          }
        },
        "required": [
          "name"
        ]
      },
      "UserDTO": {
        "type": "object",
        "description": "User data for registration",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the user",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Username for authentication",
            "example": "john_doe",
            "maxLength": 50,
            "minLength": 3
          },
          "password": {
            "type": "string",
            "description": "User password (only for registration/update)",
            "example": "password123",
            "maxLength": 100,
            "minLength": 6
          },
          "displayName": {
            "type": "string",
            "description": "Display name shown to other users",
            "example": "John Doe",
            "maxLength": 100,
            "minLength": 2
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "avatar": {
            "type": "string",
            "description": "URL to user's avatar image",
            "example": "https://example.com/avatar.jpg"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of last login",
            "example": "2024-01-15T10:30:00"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when user was created",
            "example": "2024-01-01T00:00:00"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when user was last updated",
            "example": "2024-01-15T10:30:00"
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the user account is enabled",
            "example": true
          },
          "name": {
            "type": "string",
            "description": "Legacy name field (deprecated)",
            "example": "John Doe"
          },
          "roles": {
            "type": "array",
            "description": "List of roles assigned to the user",
            "items": {
              "$ref": "#/components/schemas/RoleDTO"
            }
          }
        },
        "required": [
          "displayName",
          "password",
          "username"
        ]
      },
      "AttachmentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "filename": {
            "type": "string",
            "minLength": 1
          },
          "originalFilename": {
            "type": "string",
            "minLength": 1
          },
          "fileUrl": {
            "type": "string",
            "minLength": 1
          },
          "contentType": {
            "type": "string",
            "minLength": 1
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "taskId": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedBy": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "contentType",
          "fileSize",
          "fileUrl",
          "filename",
          "originalFilename"
        ]
      },
      "CommentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "editedAt": {
            "type": "string",
            "format": "date-time"
          },
          "mentions": {
            "type": "string"
          },
          "edited": {
            "type": "boolean"
          },
          "author": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "taskId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "content"
        ]
      },
      "LabelDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1
          },
          "color": {
            "type": "string",
            "maxLength": 7,
            "minLength": 4
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "boardId": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "color",
          "name"
        ]
      },
      "TaskDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "priority": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDTO"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentDTO"
            }
          },
          "labels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabelDTO"
            }
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "assignee": {
            "$ref": "#/components/schemas/UserDTO"
          }
        },
        "required": [
          "title"
        ]
      },
...
      "ColumnDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "type": "string"
          },
          "taskLimit": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDTO"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "BoardDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "settings": {
            "type": "string"
          },
          "invitationCode": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnDTO"
            }
          },
          "labels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabelDTO"
            }
          }
        },
        "required": [
          "name"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}